run storage tests
